// app/lib/definitions/enums/png/PngMetaField.ts
enum PngMetaField {
  nome = "nome",
  descrizione = "descrizione",
  titolo = "titolo",
  allineamento = "allineamento",
  dominioAllineamento = "dominioallineamento",
  mansione = "mansione",
  luogo = "luogo",
  fazione = "fazione",
  aspetto = "aspetto",
  personalita = "personalita",
  motivazioni = "motivazioni",
  segreti = "segreti",
}

export default PngMetaField;